Сделал авторизацию и регистрацию

+++ RetrofitProject/app/src/main/java/com/example/retrofitproject/DataClasses/SocialMapObject.kt
@@ -3,8 +3,8 @@ package com.example.retrofitproject.DataClasses
 data class SocialMapObject(
     val rating: Int,
     val id: Int,
-    val x: Int,
-    val y: Int,
+    val x: Double,
+    val y: Double,
     val display_name: String,
     val adress: String,
     val images: String,
